working with --disable-recompose flag
