+++ src/Pass/Passes.hpp
@@ -33,10 +33,12 @@ std::unique_ptr<mlir::Pass> createScrubDisposablePass(bool closeAfter = true);
 std::unique_ptr<mlir::Pass> createONNXOpTransformPass();
 std::unique_ptr<mlir::Pass> createONNXOpTransformPass(int threshold,
     bool report, bool targetCPU, bool enableSimdDataLayoutOpt,
-    bool enableConvOptPass);
+    bool enableConvOptPass, bool enableRecomposeOptPass);
 
 /// Pass for rewriting inside frontend dialect.
 std::unique_ptr<mlir::Pass> createDecomposeONNXToONNXPass(
+    const std::string &target = "");
+std::unique_ptr<mlir::Pass> createRecomposeONNXToONNXPass(
     const std::string &target = "");
 
 std::unique_ptr<mlir::Pass> createConvOptONNXToONNXPass(
@@ -67,7 +69,8 @@ std::unique_ptr<mlir::Pass> createStandardFuncReturnPass();
 
 /// Pass that combines multiple ONNX dialect transformations,
 /// including shape inference.
-std::unique_ptr<mlir::Pass> createONNXHybridTransformPass();
+std::unique_ptr<mlir::Pass> createONNXHybridTransformPass(
+    bool enableRecomposition);
 
 /// Pass for analyzing unknown dimension in ONNX operations.
 std::unique_ptr<mlir::Pass> createONNXDimAnalysisPass();
